+++ tests/foragers/test__forager.py
@@ -273,3 +273,32 @@ def test__find_rse(nbf_t: NbForager, params, expected: Any):
     else:
         with pytest.raises(expected):
             nbf_t.ipam.prefixes.find_rse(**params)
+
+
+@pytest.mark.parametrize("kwargs, expected", [
+    ({"id": [1]}, {"id": []}),  # nested=True
+    ({"id": [2]}, {"id": [2]}),  # nested=False
+    ({"id": [9]}, {"id": [9]}),  # absent
+    ({"id": [1, 2, 9]}, {"id": [2, 9]}),  # combo
+    ({"id": {1}}, {"id": {1}}),  # not list
+    ({"id": 1}, {"id": 1}),  # not list
+    ({"name": "DEVICE1"}, {"name": "DEVICE1"}),
+    ({"name": ["DEVICE1"]}, {"name": ["DEVICE1"]}),
+    ({}, {}),
+])
+def test__delete_existing_nested_ids(nbf_r: NbForager, kwargs, expected):
+    """Forager._delete_existing_nested_ids().
+
+    NbForager.tree and NbForager.root has 3 devices: DEVICE1, DEVICE2, DEVICE3.
+    DEVICE1 has: tags=TAG1, device_role=DEVICE ROLE1, serial=SERIAL1
+    DEVICE2 has: tags=TAG1, device_role=DEVICE ROLE1, serial=SERIAL2
+    DEVICE3 has: tags=TAG3, device_role=DEVICE ROLE3, serial=SERIAL1
+    """
+    nbf_r.root.dcim.devices = {k: v for k, v in nbf_r.root.dcim.devices.items() if k in [1, 2]}
+    nbf_r.root.dcim.devices[1]["_nested"] = True
+    nbf_r.root.dcim.devices[2]["_nested"] = False
+
+    nbf_r.dcim.devices._delete_existing_nested_ids(kwargs)
+
+    diff = list(dictdiffer.diff(kwargs, expected))
+    assert not diff